backend/pyproject.toml
[tool.poetry]
name = "boss-backend"
version = "0.1.0"
description = ""
authors = ["Santosh Kumar <santoshjnvropar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
django-cors-headers = "^4.6.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
pillow = "^11.1.0"
pyjwt = "^2.10.1"
python-json-logger = "^3.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
